+++ src/nsidc/metgen/metgen.py
@@ -40,7 +40,7 @@ from rich.prompt import Confirm, Prompt
 from nsidc.metgen import aws, config, constants
 from nsidc.metgen.collection_metadata import get_collection_metadata
 from nsidc.metgen.models import CollectionMetadata
-from nsidc.metgen.readers import generic, registry, utilities
+from nsidc.metgen.readers import registry, utilities
 from nsidc.metgen.spatial import create_flightline_polygon
 
 # -------------------------------------------------------------------
@@ -366,7 +366,7 @@ def process(configuration: config.Config) -> None:
             premet_filename=premet_file,
             spatial_filename=spatial_file,
             reference_data_filename=reference_data_file,
-            data_reader=data_reader(configuration.auth_id, reference_data_file),
+            data_reader=registry.lookup(Path(reference_data_file).suffix),
         )
         for name, reference_data_file, data_files, browse_files, premet_file, spatial_file in grouped_granule_files(
             configuration
@@ -376,20 +376,6 @@ def process(configuration: config.Config) -> None:
     results = [pipeline(g) for g in granules]
 
     summarize_results(results)
-
-
-def data_reader(
-    auth_id: str, data_file: str
-) -> Callable[[str, str, str, config.Config], dict]:
-    """
-    Determine which file reader to use for the given data file.
-    """
-    _, extension = os.path.splitext(data_file)
-
-    try:
-        return registry.lookup(auth_id, extension)
-    except (KeyError, Exception):
-        return generic.extract_metadata
 
 
 # -------------------------------------------------------------------