+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## UNRELEASED
+
+* Issue-255: Remove CSV reader capability (generic & SNOWEX-specific readers)
+* Fix configuration issues with some integration tests
+
+## v1.11.0 (2025-09-02)
+
+ * Minor version release of all issues in v1.11.0rc1, rc2, rc3
+
 ## v1.11.0rc3 (2025-08-25)
 
 * Issue-259: Fix white space bug in handling of `.premet` contents.